Build frontend part with react, chakra-ui and zustand

+++ backend/controller/product.controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Product from "../models/product.model.js";
 
 export const getProducts = async (req, res) => {
@@ -55,11 +56,17 @@ export const putProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Product not found" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("error in deleting product", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server error" });
   }
 };
